% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{euclidean function}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer.}

\item{b}{An integer.}
}
\value{
This function return a number that shows the greatest common divisor
  of \code{a} and \code{b}.
}
\description{
This function implements the Euclidian algorithm to find the greatest common
divisor of two numbers. The description of the algorithm with pseudocode can
be found here: \url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
\examples{
euclidean(123612, 13892347912) #4
euclidean(100, 1000) #100

}
