% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlab03.R
\docType{package}
\name{rlab03}
\alias{rlab03}
\alias{rlab03-package}
\title{rlab03 package}
\description{
This package has two functions \code{\link{euclidean}} and
\code{\link{dijkstra}}.The first function can be used to find the greatest
common divisor of two numbers and the second one takes a graph and an initial
node and calculates the shortest path from the initial node to every other
node in the graph.
}
