% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph-data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{An object of class \code{data.frame} with 18 rows and 3 columns.}
\usage{
data(wiki_graph)
}
\description{
This is wiki_graph for testing dijkstra function.
}
\references{
\url{data_blah.com}
}
\author{
Min-Chun \email{mcshihs@gmail.com}
}
\keyword{datasets}
